kconfig: fixes and sync to next-20240830
There's a slew of fixes needed to use the latest linux-next, most
of them are just the Makefile magic, and trimming that down to our
own needs. This gets us up to next-20240830 and we can now build
all targets.

We modify the default target so we can test building conf, mconf, nconf
through continuous integration later.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>