kconfig: add optional selective yaml output support
kconfig is used outside of Linux, and one of uses of Kconfig is to
also allow kconfig to be used for automation on kdevops by leveraging
a smaller subset of variables for yaml run time for ansible runs.
There is no need to clutter a full yaml file with every single config
we have as we do in the kernel, and so this lets users decide if they
want all or just a few select key symbols as part of the yaml output.

What this will do is save us the pain of doing the selective
transformation we currently do.

You can test with:

export KCONFIG_YAMLCFG=".yaml"
export KCONFIG_YAMLCFG_ALL="y"
rm -f .config .yaml
make defconfig
head -10 .yaml