Makefile: fix ordering to account for parser and lexical analyzer changes
If we modify our lexical analyzer (lexer.l) or parser (parser.y) we
end up looking for parser.tab.h on building the lexer.c, this needs
to be fixed. And so also just fix the build order so we always build
these first.

If someone runs make clean, make sure we kill their older lexical
analyzer output and output parser as we may make modifications.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>